Move several string rewrites to extended rewrites (#11509)
This is work towards eliminating nested calls to the rewriter from
within the strings theory rewriter.

Several rewrites of this form are now "demoted" to the extended
rewriter.

This is part of a branch that shows no difference in performance wrt
main on SMT-LIB and is roughly 2% faster on an industrial set of
interest.