Refactor congruence proof rules (#11501)
This refactors the congruence rules of CPC.

The motivation is:
1. We should not export "Kind" nodes to the API as these are not useful
to a user.
2. We no longer have to disambiguate operators when applying congruence.
This includes parameterized operators like bvadd, binary vs unary minus,
as well as uninterpreted functions with overloaded symbols (the latter
is currently broken in cpc output since it is not possible to tell from
cvc5's internals whether a UF was overloaded).
3. Efficiency, as the proofs contain one fewer argument and the
resulting proof checkers should be more efficient.

One question is whether we should further introduce a `BINDER_CONG` for
congruence over binders. These are currently a special case of `CONG`.

This PR enables the cpc tester for 13 regressions that could not be
checked due to congruence on overloaded UF.

---------

Co-authored-by: Haniel Barbosa <hanielbbarbosa@gmail.com>