Add option to disallow overloading of terms and sorts (#11493)
This adds an option to disallow sorts and terms that have the same name.
It is disabled by default, and thus the behavior of our parser is
preserved.

This option make cvc5 *more* pedantic than the SMT-LIB 2.X standard.
Future revisions (e.g. SMT-LIB 3.0) may disallow this, in which case
this option can be enabled.

The motivation is that any proofs we generate will fail to check e.g.
with ethos when this is the case.

There is one benchmark in SMT-LIB with this issue.